requirements.txt
cx_Oracle
pandas
Pillow
tabulate
mongo